added button & method to change all aircond statuses

+++ app/workers/ac_timer_worker.rb
@@ -6,6 +6,6 @@ class AcTimerWorker
   	arguments = args[0]
   	puts "Ac Timer Worker Running!"
   	aircond = Aircond.find(arguments["aircond_id"])
-    aircond.send_signal(status:arguments["status"])
+  	aircond.send_signal(status:arguments["status"]) if aircond.get_state[:status] != arguments["status"]
   end
 end
